+++ tests/kgpgimport.cpp
@@ -24,7 +24,7 @@ void KGpgImportTest::testImportTextKey()
 		QCOMPARE(transaction->getImportedKeys().size(), 1);
 		QString keyID = QLatin1String("BA7695F3C550DF14");
 		QVERIFY(transaction->getImportedKeys().first().startsWith(keyID));
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
@@ -41,7 +41,7 @@ void KGpgImportTest::testImportKeyFromFile()
 		QCOMPARE(transaction->getImportedKeys().size(), 1);
 		QString keyID = QLatin1String("BA7695F3C550DF14");
 		QVERIFY(transaction->getImportedKeys().first().startsWith(keyID));
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
@@ -63,7 +63,7 @@ void KGpgImportTest::testImportSameKeyTwice()
 		QVERIFY(transaction->getImportedKeys().first().startsWith(keyID));
 		QVERIFY(transaction->getMessages().join(QLatin1Char(' ')).contains(duplicateMsg));
 		QEXPECT_FAIL("", "Test is broken. Possible bug!", Continue);
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
@@ -79,7 +79,7 @@ void KGpgImportTest::testImportIdsAll()
 		QCOMPARE(transaction->getImportedIds(-1).size(), 1);
 		QString fingerprint = QLatin1String("FBAF08DD7D9D0921C15DDA9FBA7695F3C550DF14");
 		QVERIFY(transaction->getImportedIds(-1).first().compare(fingerprint) == 0);
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
@@ -96,7 +96,7 @@ void KGpgImportTest::testImportIdsUnchanged()
 		QCOMPARE(transaction->getImportedIds(0).size(), 1);
 		QString fingerprint = QLatin1String("FBAF08DD7D9D0921C15DDA9FBA7695F3C550DF14");
 		QVERIFY(transaction->getImportedIds(0).first().compare(fingerprint) == 0);
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
@@ -111,7 +111,7 @@ void KGpgImportTest::testLogMessage()
 	QObject::connect(transaction, &KGpgImport::done, [transaction](int result) {
 		QString msg = QLatin1String("IMPORT_OK 1 FBAF08DD7D9D0921C15DDA9FBA7695F3C550DF14");
 		QVERIFY(transaction->getMessages().join(QLatin1Char(' ')).contains(msg));
-		QCOMPARE(result, KGpgTransaction::TS_OK);
+		QCOMPARE(result, static_cast<int>(KGpgTransaction::TS_OK));
 	});
 	transaction->start();
 	QVERIFY(spy.wait());
